[TASK] Basic implementation
Please call
https://jungfrau.dev/?id=1&eID=routing&route=translationapi/<extension-key>

to get labels from the corresponding locallang.xlf file. You may add a segment
to the route to filter by a given prefix, e.g., "/module" for every label
whose key starts with "module." (minds the period).

To get labels in another language, just do it like a regular TYPO3 page, that
is pass a parameter "L" to the URI.